Simplify ESLint config to avoid missing packages
