--- src/components/ui/use-toast.ts
@@ -1,3 +1,0 @@
-import { useToast, toast } from "@/hooks/use-toast";
-
-export { useToast, toast };
