+++ src/hooks/useCrisisSystem.ts
@@ -4,7 +4,7 @@ import { toast } from 'sonner';
 import { useAuth } from '@/contexts/AuthContext';
 import { voiceActivationService } from '@/services/voiceActivationService';
 import { escalateCrisis } from '@/services/crisisEscalationService';
-import { useAuditLogger } from '@/hooks/useAuditLogger';
+import { useSecureAuditLogger } from '@/hooks/useSecureAuditLogger';
 
 type RiskLevel = 'low' | 'moderate' | 'high' | 'severe';
 
@@ -28,7 +28,7 @@ export const useCrisisSystem = () => {
   const [voiceListening, setVoiceListening] = useState(false);
   const [hasLocationPermission, setHasLocationPermission] = useState(false);
   const { user } = useAuth();
-  const { log } = useAuditLogger();
+  const { log } = useSecureAuditLogger();
 
   useEffect(() => {
     // Initialize voice activation if supported
